+++ fixtures/protocol/htty/executables/abort_after_bootstrap.rb
@@ -7,4 +7,4 @@ require "protocol/htty"
 
 $stdout.binmode
 
-Protocol::HTTY::Stream.new($stdout).write_bootstrap
+Protocol::HTTY::Stream.new($stdin, $stdout).write_bootstrap